requirements.txt
python_dotenv
requests
langchain